feat: rename energy flame to flame ember
