+++ components/ui/index.ts
@@ -20,7 +20,7 @@ export { Fab } from "./Fab";
 export { CreateSheet } from "./CreateSheet";
 export { QuickPeek } from "./QuickPeek";
 export { MobileNav } from "../mobile-nav";
-export { EnergyFlame } from "./EnergyFlame";
+export { FlameEmber } from "./FlameEmber";
 
 // New UX components
 export {
